Update inserting_data.sql
Earlier we had place_num in the Database which was a Number attribute and we had difficulties checking the duplication and uniqueness in Residence_Hall and Student_Flat. Now from Dr. Shafaq’s feedback in phase 1 presentation, we replaced the number attribute  to varchar and now we can check it easily.